Integration tests refactoring and new test (#887)
* Moved some classes from the former NodeBuilder.cs to other files because the file was getting huge. Nothing else was changed.

* Added test for miner-connector scenario described in PR #879.

* Fixes based on @mikedennis review.
