+++ src/Stratis.Bitcoin.IntegrationTests/EnvironmentMockUpHelpers/NodeRunners.cs
@@ -1,0 +1,252 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Text;
+using NBitcoin;
+using NBitcoin.Protocol;
+using Stratis.Bitcoin.Builder;
+using Stratis.Bitcoin.Configuration;
+using Stratis.Bitcoin.Features.BlockStore;
+using Stratis.Bitcoin.Features.Consensus;
+using Stratis.Bitcoin.Features.MemoryPool;
+using Stratis.Bitcoin.Features.Miner;
+using Stratis.Bitcoin.Features.RPC;
+using Stratis.Bitcoin.Features.Wallet;
+using Stratis.Bitcoin.Utilities;
+
+namespace Stratis.Bitcoin.IntegrationTests
+{
+    public class BitcoinCoreRunner : INodeRunner
+    {
+        private string bitcoinD;
+
+        public BitcoinCoreRunner(string bitcoinD)
+        {
+            this.bitcoinD = bitcoinD;
+        }
+
+        private Process process;
+
+        public bool HasExited
+        {
+            get { return this.process == null && this.process.HasExited; }
+        }
+
+        public void Kill()
+        {
+            if (!this.HasExited)
+            {
+                this.process.Kill();
+                this.process.WaitForExit();
+            }
+        }
+
+        public void Start(string dataDir)
+        {
+            this.process = Process.Start(new FileInfo(this.bitcoinD).FullName,
+                $"-conf=bitcoin.conf -datadir={dataDir} -debug=net");
+        }
+    }
+
+    public class StratisBitcoinPosRunner : INodeRunner
+    {
+        private Action<IFullNodeBuilder> callback;
+
+        public StratisBitcoinPosRunner(Action<IFullNodeBuilder> callback = null) : base()
+        {
+            this.callback = callback;
+        }
+
+        public bool HasExited
+        {
+            get { return this.FullNode.HasExited; }
+        }
+
+        public void Kill()
+        {
+            if (this.FullNode != null)
+            {
+                this.FullNode.Dispose();
+            }
+        }
+
+        public void Start(string dataDir)
+        {
+            NodeSettings nodeSettings = new NodeSettings("stratis", InitStratisRegTest(), ProtocolVersion.ALT_PROTOCOL_VERSION).LoadArguments(new string[] { "-conf=stratis.conf", "-datadir=" + dataDir });
+
+            var node = BuildFullNode(nodeSettings, this.callback);
+
+            this.FullNode = node;
+            this.FullNode.Start();
+        }
+
+        public static FullNode BuildFullNode(NodeSettings args, Action<IFullNodeBuilder> callback = null)
+        {
+            FullNode node;
+
+            if (callback != null)
+            {
+                var builder = new FullNodeBuilder().UseNodeSettings(args);
+
+                callback(builder);
+
+                node = (FullNode)builder.Build();
+            }
+            else
+            {
+                node = (FullNode)new FullNodeBuilder()
+                    .UseNodeSettings(args)
+                    .UseStratisConsensus()
+                    .UseBlockStore()
+                    .UseMempool()
+                    .UseWallet()
+                    .AddPowPosMining()
+                    .AddRPC()
+                    .Build();
+            }
+
+            return node;
+        }
+
+        public FullNode FullNode;
+
+        private static Network InitStratisRegTest()
+        {
+            // TODO: move this to Networks
+            var net = Network.GetNetwork("StratisRegTest");
+            if (net != null)
+                return net;
+
+            Block.BlockSignature = true;
+            Transaction.TimeStamp = true;
+
+            var consensus = Network.StratisTest.Consensus.Clone();
+            consensus.PowLimit = new Target(uint256.Parse("7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"));
+
+            consensus.PowAllowMinDifficultyBlocks = true;
+            consensus.PowNoRetargeting = true;
+
+            // The message start string is designed to be unlikely to occur in normal data.
+            // The characters are rarely used upper ASCII, not valid as UTF-8, and produce
+            // a large 4-byte int at any alignment.
+            var pchMessageStart = new byte[4];
+            pchMessageStart[0] = 0xcd;
+            pchMessageStart[1] = 0xf2;
+            pchMessageStart[2] = 0xc0;
+            pchMessageStart[3] = 0xef;
+            var magic = BitConverter.ToUInt32(pchMessageStart, 0); //0x5223570;
+
+            var genesis = Network.StratisMain.GetGenesis().Clone();
+            genesis.Header.Time = 1494909211;
+            genesis.Header.Nonce = 2433759;
+            genesis.Header.Bits = consensus.PowLimit;
+            consensus.HashGenesisBlock = genesis.GetHash();
+
+            Guard.Assert(consensus.HashGenesisBlock == uint256.Parse("0x93925104d664314f581bc7ecb7b4bad07bcfabd1cfce4256dbd2faddcf53bd1f"));
+
+            var builder = new NetworkBuilder()
+                .SetName("StratisRegTest")
+                .SetConsensus(consensus)
+                .SetMagic(magic)
+                .SetGenesis(genesis)
+                .SetPort(18444)
+                .SetRPCPort(18442)
+                .SetBase58Bytes(Base58Type.PUBKEY_ADDRESS, new byte[] { (65) })
+                .SetBase58Bytes(Base58Type.SCRIPT_ADDRESS, new byte[] { (196) })
+                .SetBase58Bytes(Base58Type.SECRET_KEY, new byte[] { (65 + 128) })
+                .SetBase58Bytes(Base58Type.ENCRYPTED_SECRET_KEY_NO_EC, new byte[] { 0x01, 0x42 })
+                .SetBase58Bytes(Base58Type.ENCRYPTED_SECRET_KEY_EC, new byte[] { 0x01, 0x43 })
+                .SetBase58Bytes(Base58Type.EXT_PUBLIC_KEY, new byte[] { (0x04), (0x88), (0xB2), (0x1E) })
+                .SetBase58Bytes(Base58Type.EXT_SECRET_KEY, new byte[] { (0x04), (0x88), (0xAD), (0xE4) });
+
+            return builder.BuildAndRegister();
+        }
+
+        /// <summary>
+        /// Builds a node with POS miner and RPC enabled.
+        /// </summary>
+        /// <param name="dir">Data directory that the node should use.</param>
+        /// <returns>Interface to the newly built node.</returns>
+        /// <remarks>Currently the node built here does not actually stake as it has no coins in the wallet,
+        /// but all the features required for it are enabled.</remarks>
+        public static IFullNode BuildStakingNode(string dir, bool staking = true)
+        {
+            NodeSettings nodeSettings = new NodeSettings().LoadArguments(new string[] { $"-datadir={dir}", $"-stake={(staking ? 1 : 0)}", "-walletname=dummy", "-walletpassword=dummy" });
+            var fullNodeBuilder = new FullNodeBuilder(nodeSettings);
+            IFullNode fullNode = fullNodeBuilder
+                .UseStratisConsensus()
+                .UseBlockStore()
+                .UseMempool()
+                .UseWallet()
+                .AddPowPosMining()
+                .AddRPC()
+                .Build();
+
+            return fullNode;
+        }
+    }
+
+    public class StratisBitcoinPowRunner : INodeRunner
+    {
+        private Action<IFullNodeBuilder> callback;
+
+        public StratisBitcoinPowRunner(Action<IFullNodeBuilder> callback = null) : base()
+        {
+            this.callback = callback;
+        }
+
+        public bool HasExited
+        {
+            get { return this.FullNode.HasExited; }
+        }
+
+        public void Kill()
+        {
+            if (this.FullNode != null)
+            {
+                this.FullNode.Dispose();
+            }
+        }
+
+        public void Start(string dataDir)
+        {
+            NodeSettings nodeSettings = new NodeSettings().LoadArguments(new string[] { "-conf=bitcoin.conf", "-datadir=" + dataDir });
+
+            var node = BuildFullNode(nodeSettings, this.callback);
+
+            this.FullNode = node;
+            this.FullNode.Start();
+        }
+
+        public static FullNode BuildFullNode(NodeSettings args, Action<IFullNodeBuilder> callback = null)
+        {
+            FullNode node;
+
+            if (callback != null)
+            {
+                var builder = new FullNodeBuilder().UseNodeSettings(args);
+
+                callback(builder);
+
+                node = (FullNode)builder.Build();
+            }
+            else
+            {
+                node = (FullNode)new FullNodeBuilder()
+                    .UseNodeSettings(args)
+                    .UseConsensus()
+                    .UseBlockStore()
+                    .UseMempool()
+                    .AddMining()
+                    .UseWallet()
+                    .AddRPC()
+                    .Build();
+            }
+
+            return node;
+        }
+
+        public FullNode FullNode;
+    }
+}
